Improved Fox-Glynn for small numbers + int overflow bugfix (Vojta).
git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@1926 bbc10eb1-c90d-0410-af57-cb519fbb1720

+++ prism/include/prism.h
@@ -36,8 +36,8 @@
 // Fox-Glynn wieghts struct
 typedef struct FoxGlynnWeights
 {
-	int left;
-	int right;
+	long left;
+	long right;
 	double total_weight;
 	double *weights;
 } FoxGlynnWeights;
